// jpt-intern-2026/frontend/src/api/models/ProjectStatus.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export enum ProjectStatus {
    DRAFT = 'DRAFT',
    PENDING_DEPT = 'PENDING_DEPT',
    PENDING_HQ = 'PENDING_HQ',
    APPROVED = 'APPROVED',
    IN_PROGRESS = 'IN_PROGRESS',
    COMPLETED = 'COMPLETED',
    REJECTED = 'REJECTED',
}