// jpt-intern-2026/frontend/src/api/models/UserResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { UserRole } from './UserRole';
export type UserResponse = {
    id: number;
    name: string;
    email: string;
    role: UserRole;
    department_id?: (number | null);
};